refactor(cpuid): make msrs_to_save_by_cpuid shared
MSR dependencies to CPUID feature flags we have
at the moment are all shared between Intel and AMD.
